+++ src/exceptions/interfaces/index.ts
@@ -1,0 +1,2 @@
+export * from './error-response.interface';
+export * from './validation-error-detail.interface';