+++ src/exceptions/interfaces/validation-error-detail.interface.ts
@@ -1,0 +1,7 @@
+import { ValidationErrorType } from '../reference/validation-error-type.reference';
+
+export interface IValidationErrorDetails {
+  type: ValidationErrorType;
+  path: string;
+  message: string;
+}