+++ src/exceptions/reference/validation-error-type.reference.ts
@@ -1,0 +1,6 @@
+export enum ValidationErrorType {
+  AlreadyExists = 'already_exists',
+  Invalid = 'invalid',
+  MissingField = 'missing_field',
+  Unrecognized = 'unrecognized',
+}