+++ src/exceptions/transformers/index.ts
@@ -1,0 +1,2 @@
+export * from './nestjs.exception-transformer';
+export * from './typeorm.exception-transformer';